fastapi/requirements.txt
fastapi[standard]
uvicorn
pydantic-settings
mariadb
sqlmodel
sqlalchemy
jproperties
pyjwt
typing
pytest
httpx
python-multipart
cryptography